requirements.txt
fastapi==0.100.0
pydantic==1.10.9
firebase-admin==6.1.0
python-dotenv==1.0.0
uvicorn==0.23.0